requirements.txt
black
isort
PyAutoGui
numpy
screeninfo
pillow
